Further clean up and extension of comments.
