+++ src/main/scala/org/remotefutures/core/Settings.scala
@@ -31,6 +31,9 @@ object Settings {
   }
 }
 
+// ==============================================================================
+// Unsure, HOW to do it reasonable. Inheritence is a very limited solution ....
+// ==============================================================================
 
 //trait AkkaOnlySettings
 //
