+++ src/main/scala/org/remotefutures/core/impl/executor/LocalRunningRemoteExecutor.scala
@@ -6,27 +6,27 @@ import scala.util.{Failure, Success}
 import scala.util.control.NonFatal
 
 /**
- * A dummy remote executor which runs the function body in a Runnable on the local ExecutionContext
+ * An remote executor which runs the function fnc in a Runnable on the local ExecutionContext
  */
-class LocalRunnableRemoteExecutor extends RemoteExecutor {
+class LocalRunningRemoteExecutor extends RemoteExecutor {
 
-  override def execute[C, T](body: () => T, bodyContext: C, promise: Promise[T]): Unit = {
+  override def execute[C, T](fnc: () => T, fncContext: C, promise: Promise[T]): Unit = {
     println("This is execute of LocalRunnableRemoteExecutor.")
 
     import scala.concurrent.ExecutionContext.Implicits.global
 
     implicit val executor: ExecutionContext = global
 
-    val runnable = new PromiseCompletingRunnable(body, promise)
+    val runnable = new PromiseCompletingRunnable(fnc, promise)
     executor.prepare.execute(runnable)
   }
 
   /**
-   * A runnable wrapping the execution of body, whose result is put the a promise.
+   * A runnable wrapping the execution of fnc, whose result is put the a promise.
    *
-   * Shamlessly copied from concurrent.impl.Future and modified.
+   * Shamelessly copied from concurrent.impl.Future and modified.
    *
-   * @param body the body to execute
+   * @param body the fnc to execute
    * @param promise is the promise to put the result of type T into
    * @tparam T return type of this distributed future.
    */
@@ -34,9 +34,9 @@ class LocalRunnableRemoteExecutor extends RemoteExecutor {
     override def run() = {
       promise complete {
         try {
-          println("Before execution of body on " + this)
-          val result: T = body() // execute body
-          println("After execution of body on " + this)
+          println("Before execution of fnc on " + this)
+          val result: T = body() // execute fnc
+          println("After execution of fnc on " + this)
           Success(result)
 
         } catch {
